#342 create BridgeUnit.explicit_label_map
